+++ Assets/_Game/Scripts/Ui/UiPlayAgain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4ea4bc095a6b4b088ec8f30a65b30cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
